requirements.txt
pandas==2.2.3
pytest==8.3.4
urllib3==2.3.0
tldextract==5.1.3
python-whois==0.9.5
tqdm==4.67.1
tabulate==0.9.0
colorama==0.4.6
